Renaming deprecated
