+++ applications/convection_diffusion_application/python_scripts/convection_diffusion_solver.py
