+++ applications/convection_diffusion_application/python_scripts/eulerian_convection_diffusion_solver.py
