+++ applications/convection_diffusion_application/python_scripts/pfem2_convection_diffusion_solver.py
