+++ applications/convection_diffusion_application/python_scripts/pure_convection_solver.py
